requirements.txt
Flask==2.2.2
Flask-Cors==3.0.10
Flask-Limiter==3.5.0
redis==5.0.1
psycopg2-binary==2.9.9
marshmallow==3.20.1
PyJWT==2.8.0
cryptography==41.0.7
numpy==1.26.2
pandas==2.1.4
scikit-learn==1.3.2
llama-cpp-python==0.2.20
python-dotenv==1.0.0
